move code into lib.rs
